requirement.txt
Pillow==8.4.0
pypng==0.0.21
PyQRCode==1.2.1
qrcode==7.3.1
wifi-qrcode-generator==0.1
